backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
chromadb==0.5.20
sentence-transformers==3.3.0
transformers==4.47.0
peft==0.13.2
bitsandbytes==0.44.1
accelerate==1.1.1
datasets==3.1.0
trl==0.12.1
ragas==0.2.6
pypdf==5.1.0
python-dotenv==1.0.1
torch==2.9.0
huggingface-hub==0.26.2
tokenizers>=0.21.0
